+++ playground/src/router/routes/index.ts
@@ -1,0 +1,15 @@
+import { modelsRoutes } from './models'
+import { cameraRoutes } from './cameras'
+import { eventsRoutes } from './events'
+import { basicRoutes } from './basic'
+import { perfRoutes } from './performance'
+import { miscRoutes } from './misc'
+
+export {
+  basicRoutes,
+  perfRoutes,
+  eventsRoutes,
+  cameraRoutes,
+  modelsRoutes,
+  miscRoutes,
+}